bug; Fixed missing lights
