+++ app/api/db-test/route.ts
@@ -6,6 +6,6 @@ export async function GET() {
     const result = await pool.query("SELECT 1 AS test")
     return NextResponse.json(result.rows)
   } catch (error) {
-    return NextResponse.json({ error: error.message }, { status: 500 })
+    return NextResponse.json({ error: error }, { status: 500 })
   }
 }
